First draft of optimal stopping
added a new method, which is to start using yuor second best observation as your threshold past a certain point (called ssc in the code). This reduces your odds of finding the very best candidate, however it (would seem it) increases the average quality of the candidate you select. A quick explaratory analysis seemed to indicate that having ssc set to anything from 0 to 0.4 performed best. Given some observed spikes, it would  seem that 0.37 could very well be the sweet spot for this value as well. An exploration via simulation will be undertaken to ascertain which looking set size gives the best average candidate when using the second best observation from the looking set as our threshold instead